+++ app/src/main/java/woowacourse/shopping/data/service/cart/CartRetrofitApiGenerator.kt
@@ -1,16 +1,16 @@
-package woowacourse.shopping.data.service
+package woowacourse.shopping.data.service.cart
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import woowacourse.shopping.user.ServerInfo
 
-object CartProductServiceGenerator {
+object CartRetrofitApiGenerator {
     private val retrofit: Retrofit =
         Retrofit
             .Builder()
             .baseUrl(ServerInfo.url)
             .addConverterFactory(GsonConverterFactory.create())
             .build()
-    val service: CartProductRetrofitApi =
-        retrofit.create(CartProductRetrofitApi::class.java)
+    val service: CartRetrofitApi =
+        retrofit.create(CartRetrofitApi::class.java)
 }
