+++ app/src/main/java/woowacourse/shopping/feature/detail/dialog/CounterDialog.kt
@@ -10,7 +10,7 @@ import androidx.fragment.app.DialogFragment
 import woowacourse.shopping.R
 import woowacourse.shopping.common_ui.CounterView
 import woowacourse.shopping.data.repository.local.CartRepositoryImpl
-import woowacourse.shopping.data.service.CartProductRemoteService
+import woowacourse.shopping.data.service.cart.CartRemoteService
 import woowacourse.shopping.databinding.DialogCounterBinding
 import woowacourse.shopping.model.ProductUiModel
 import woowacourse.shopping.util.getParcelableCompat
@@ -52,7 +52,7 @@ class CounterDialog : DialogFragment(), CounterDialogContract.View {
     private fun setInitPresenter(product: ProductUiModel, cartId: Long?) {
         presenter = CounterDialogPresenter(
             this,
-            CartRepositoryImpl(CartProductRemoteService()),
+            CartRepositoryImpl(CartRemoteService()),
             product,
             cartId,
         )
